Score and Reset
Added a Score file that displays a new score screen. The screen shows
the stats of the user who just took the quiz and has options to leave,
go to the main menu, or retake the quiz. Also added a reset file for
quick resetting. Also cleaned up a bit of pointless lines.